% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{fars_map_state}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{numeric code representing state}

\item{year}{year of interest}
}
\value{
a state map with the locations of accidents marked as dots
}
\description{
plot the location of accidents on any state map. If the state code is
invalid, an error message will indicate that. No plot will be generated if
there is no accident for the state of interest during the year(s) of
interest. Instead, a message will show "no accident to plot". maps and dplyr
need to be imported for this function.
}
\examples{

\dontrun{
fars_map_state(12, 2015)
}

}
