% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{fars_read}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{the name of file that contains the data of interest. the file
name con be contructed using another function make_filename in the same
package}
}
\value{
data in the format of data_frame
}
\description{
read the file into R by file name.
}
\note{
Where the working directory does not contain the furnished file name,
  an error meesage stating that file doesnot exist will appear. this function
  is for internal use only, not exported.
}
\examples{

\dontrun{
filename <- make_filename(2015)
data_2015 <- fars_read(filename)
}
}
