% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{fars_read_years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{years of interest, it could be either a single year of a vector
of years}
}
\value{
a list of data, each element is a data_frame for the data for a given
  year
}
\description{
Read data files according to the years of interest.  dplyr needs to be imported for this function.
}
\note{
If the data for specified years of interest are not present, a warning message will show
those years as invalid year.
}
\examples{

\dontrun{
data_2013 <- fars_read_years(2013)

data_2013to2015 <- fars_read_years(c(2013, 2014, 2015))
}

}
