% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{fars_summarize_years}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{the years of interest. It could be a single year or a vector of
years}
}
\value{
summary of monthly tally of the number of accidents for the years of
  interest
}
\description{
summarize the monthly total of the accidents for year(s) of interest. dplyr
and tidyr need to be imported for this function.
}
\examples{

\dontrun{

sum_2015 <- fars_summarize_years(2015)
sum_2013to2015 <- fars_summarize_years(c(2013, 2014, 2015))

}

}
