% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{make_filename}
\usage{
make_filename(year)
}
\arguments{
\item{year}{year as numeric or integer, even a string year would work}
}
\value{
file name of the data for the year of interest
}
\description{
construct a file name according to the year of interest
}
\note{
for internal use only, not exported
}
\examples{

\dontrun{
filename <- make_filename(2015)
filename2 <- make_filename("2013")
}

}
